requirements.txt
selenium==4.15.0
pytest==7.4.3
allure-pytest==2.13.2
webdriver-manager==4.0.1
pytest-html==4.1.1
python-dotenv==1.0.0
requests==2.31.0
faker==20.1.0